src/requirements.txt
a2wsgi==1.10.10
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.16.5
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
argcomplete==3.6.2
asgiref==3.9.1
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
build==1.3.0
cachetools==5.5.2
cadwyn==5.4.4
cattrs==25.2.0
certifi==2025.8.3
cffi==2.0.0
chardet==3.0.4
charset-normalizer==3.4.3
chromadb==1.0.20
click==8.2.1
cloudpickle==3.1.1
coloredlogs==15.0.1
colorlog==6.9.0
config==0.5.1
configparser==7.2.0
contourpy==1.3.3
cron_descriptor==2.0.6
croniter==6.0.0
cryptography==41.0.7
cycler==0.12.1
databricks-sdk==0.65.0
db-dtypes==1.4.3
deep-translator==1.11.4
Deprecated==1.2.18
dill==0.4.0
distro==1.9.0
dnspython==2.8.0
docker==7.1.0
docstring_parser==0.17.0
durationpy==0.10
email-validator==2.3.0
fastapi==0.116.1
fastapi-cli==0.0.11
fastapi-cloud-cli==0.1.5
feedparser==6.0.11
filelock==3.19.1
Flask==3.1.2
flatbuffers==25.2.10
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.7.0
gcloud-aio-auth==4.2.3
gcloud-aio-bigquery==7.1.0
gcloud-aio-storage==9.2.0
gitdb==4.0.12
GitPython==3.1.45
google-ads==28.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-aiplatform==1.112.0
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-automl==2.16.4
google-cloud-bigquery==3.37.0
google-cloud-bigquery-datatransfer==3.19.2
google-cloud-bigtable==2.32.0
google-cloud-build==3.32.0
google-cloud-compute==1.37.0
google-cloud-container==2.59.0
google-cloud-core==2.4.3
google-cloud-datacatalog==3.27.1
google-cloud-dataflow-client==0.9.0
google-cloud-dataform==0.6.2
google-cloud-dataplex==2.11.0
google-cloud-dataproc==5.21.0
google-cloud-dataproc-metastore==1.19.0
google-cloud-dlp==3.31.0
google-cloud-kms==3.5.1
google-cloud-language==2.17.2
google-cloud-logging==3.12.1
google-cloud-memcache==1.12.2
google-cloud-monitoring==2.27.2
google-cloud-orchestration-airflow==1.17.5
google-cloud-os-login==2.17.2
google-cloud-pubsub==2.31.1
google-cloud-redis==2.18.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-spanner==3.57.0
google-cloud-speech==2.33.0
google-cloud-storage==2.19.0
google-cloud-tasks==2.19.3
google-cloud-texttospeech==2.29.0
google-cloud-translate==3.21.1
google-cloud-videointelligence==2.16.2
google-cloud-vision==3.10.2
google-cloud-workflows==1.18.2
google-crc32c==1.7.1
google-genai==1.38.0
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.4
grpc-google-iam-v1==0.14.2
grpc-interceptor==0.15.4
grpcio==1.74.0
grpcio-gcp==0.2.2
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
h2==3.2.0
hf-xet==1.1.8
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
json-merge-patch==0.3.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
kubernetes==33.1.0
lazy-object-proxy==1.12.0
libcst==1.8.4
linkify-it-py==2.0.3
lockfile==0.12.2
looker-sdk==25.10.0
lxml==6.0.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
methodtools==0.4.7
mmh3==5.2.0
more-itertools==10.8.0
mpmath==1.3.0
msgspec==0.19.0
multidict==6.6.4
narwhals==2.1.2
networkx==3.5
numpy==2.3.2
oauthlib==3.3.1
onnxruntime==1.22.1
openai==1.100.2
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.11.2
overrides==7.7.0
packaging==25.0
pandas==2.3.2
pandas-gbq==0.29.2
pathspec==0.12.1
pendulum==3.1.0
pillow==11.3.0
pluggy==1.6.0
posthog==5.4.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
psycopg2-binary==2.9.10
pyarrow==21.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybase64==1.4.2
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
pydata-google-auth==1.9.1
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
PyPika==0.48.9
pyproject_hooks==1.2.0
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
PyYAML-ft==8.0.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-oauthlib==2.0.0
retryhttp==1.3.3
rfc3986==1.5.0
rich==14.1.0
rich-argparse==1.7.1
rich-toolkit==0.15.1
rignore==0.6.4
rpds-py==0.27.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.1.0
sentry-sdk==2.37.1
setproctitle==1.3.7
setuptools==66.1.1
sgmllib3k==1.0.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==1.4.54
sqlalchemy-bigquery==1.15.0
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.42.0
sqlparse==0.5.3
starlette==0.47.2
streamlit==1.48.1
structlog==25.4.0
svcs==25.1.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
text-unidecode==1.3
threadpoolctl==3.6.0
tokenizers==0.21.4
toml==0.10.2
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.55.2
typer==0.16.1
types-requests==2.32.4.20250809
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uc-micro-py==1.0.3
universal_pathlib==0.2.6
uritemplate==4.2.0
urllib3==2.5.0
uuid6==2025.0.1
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wirerope==1.0.0
wrapt==1.17.3
yarl==1.20.1
zipp==3.23.0
